+++ CHANGES.rst
@@ -4,8 +4,8 @@ Products.MeetingLiege Changelog
 4.1rc2 (unreleased)
 -------------------
 
-- Nothing changed yet.
-
+- Fixed migration, while migrating MeetingCategory.groupsOfMatter, consider every categories (getCategories(onlySelectable=False), or some
+  categories end not migrated.
 
 4.1rc1 (2019-06-14)
 -------------------